Add missing unsigned long casts to size_t prints
This is a continuation of 07b64dc51a19386a. In the aforementioned commit
I accidentally missed one cast in rnd_stack_print implementation, and I
completely omitted the testing unit.

Compilation and testing has been proven successful on a 32-bit i686
Intel Atom processor. No warnings or suspicious behavior, everything
seems to work well.

+++ src/rnd_stack.c
@@ -2987,7 +2987,7 @@ int rnd_stack_print(struct rnd_stack *stack)
 	}
 #endif
 	printf("rnd_stack_print()\nsize/capacity: %lu/%lu, elem_size: %lu\n",
-		(unsigned long)stack->size, (unsigned long)stack->capacity, stack->elem_size);
+		(unsigned long)stack->size, (unsigned long)stack->capacity, (unsigned long)stack->elem_size);
 	for (i = stack->size; i-- > 0;) {
 		const void *const elem = (char*)stack->data + i * stack->elem_size;
 		printf("[%lu]\t%p\n", (unsigned long)stack->size - 1 - i, elem);
